[DotNetCore] Prevent alternative guid project used for new projects
The alternative project type guid definition did not specify an
alias so C# was used by default. Since this DotNetProjectType was
inserted before the main C# one it could be used when creating a
new project since a match would be made of C# against the
DotNetProjectType's TypeAlias. To avoid this a unique alias is
defined. This fixes the packaging tests failing with:

1) CreateMultiPlatformProjectFromTemplateWithAndroidOnly
(MonoDevelop.Packaging.Tests.ProjectTemplateTests.CreateMultiPlatformProjectFromTemplateWithAndroidOnly)
     Expected: not null
  But was:  null

  at MonoDevelop.Packaging.Tests.ProjectTemplateTests+<CreateMultiPlatformProjectFromTemplateWithAndroidOnly>d__2.MoveNext ()