[Core] Dispose HttpClient response in FileService
Ensure the HttpResponseMessage is disposed. Also removed an exception
handler that was looking for the not modified (304) status code which
is no longer valid since using the HttpClient an exception is no longer
thrown when this status is returned. Also make sure the content
is not loaded into memory since it can be copied directly to a file
by using HttpCompletionOption.ResponseHeadersRead option which
will return the response when the headers are read without reading
the content.