[Ide] ImageLoader improve memory use when using HttpClient
When requesting the image from a url the image is no longer loaded
completely into memory in the HttpRequestMessage's content. Now
the HttpCompletionOption.ResponseHeadersRead is used and the content
is then copied directly to a file. Also made sure the response is
always disposed. Previously it was only disposed when a file was
downloaded.