[NuGet] Support using native HttpMessageHandlers
Use HttpMessageHandler from MonoDevelop.Core's HttpClientProvider.
NuGet has custom http source authentication, since a NuGet.Config
file can supply credentials, so the HttpClientProvider supports a
SourceAuthenticationRequired property on the HttpClientSettings to
avoid two http source authentication handlers being created.
Added a custom HttpHandlerResourceV3Provider which will use the
HttpMessageHandler from the HttpClientProvider.

Note that the HttpHandlerResource has a HttpClientHandler property
which the NSUrlSessionHandler cannot supply so a dummy instance of
the HttpClientHandler is used if one cannot be found in the
HttpMessageHandler chain.