[NuGet] Fix credentials not accepted with package sources
When prompted for credentials to access a password protected package
source the credentials were not being sent to the server. The problem
was that the credentials need to be used with the HttpClientHandler's
Credentials which was not being done. Added a new interface
IHttpCredentialsHandler which needs to be implemented by one of
the HttpMessageHandlers so that credentials can be set by the
HttpSourceAuthenticationHandler when the user enters the username
and password.

Note that this is not working with the NSUrlSessionHandler currently.
The managed default handler is working as expected.