[NuGet] Fix multiple credential dialogs after cancelling the dialog
When the NSUrlSessionHandler is used and a package source needs
credentials which are not stored in the NuGet.Config file the
credential dialog will open when the package source is used. Clicking
the Cancel button would not keep the dialog closed for the current
NuGet action. The request would be retried and the dialog would be
displayed multiple times. The managed http client handler did not
have this problem. To fix this the original NuGet version of the
HttpSourceAuthenticationHandler is now used by the NuGet addin. This
has extra logic based on the current ActivityCorrelationId which
will prevent the dialog from being displayed multiple times.