[NuGet] Do not prompt for proxy credentials in preferences
When checking the package sources in preferences if a proxy is
required and the credentials are not set then every package source
url would trigger the credentials dialog to be displayed. Now the
ProxyAuthenticationHandler can be made non-interactive to avoid this.
This also prevents the Add Packages dialog from prompting for
credentials when attempting to download images that use http instead
of https. The image downloads now do not prompt which works around
this mono problem - https://github.com/mono/mono/issues/10622