[Core] Evaluate multi-valued MSBuild search path
MSBuild, that is included with Mono 5.16, has an MSBuild.dll.config
file that contains search path values that are multi-valued and
also need evaluating.

<searchPaths os="osx">
  <property name="MSBuildExtensionsPath" value="$(MSBuildExtensionsPathFallbackPathsOverride);/Library/Frameworks/Mono.framework/External/xbuild/" />
  <property name="MSBuildExtensionsPath32" value="$(MSBuildExtensionsPathFallbackPathsOverride);/Library/Frameworks/Mono.framework/External/xbuild/" />
  <property name="MSBuildExtensionsPath64" value="$(MSBuildExtensionsPathFallbackPathsOverride);/Library/Frameworks/Mono.framework/External/xbuild/" />
  <property name="VSToolsPath" value="/Library/Frameworks/Mono.framework/External/xbuild/Microsoft/VisualStudio/v$(VisualStudioVersion)" />
</searchPaths>

The property/values are used when looking for MSBuild imports. The
value was being used as the full path without being evaluated and
without being split into two paths. This resulted in imports that
are in the External/xbuild directory, such as Xamarin.iOS, to not
be found.

Fixes VSTS #682034 - "Tabbed Xamarin Forms App" fails to build due to
error "Your project.json doesn't have a runtimes section. You should
add '"runtimes": { "win": { } }' to your project.json and then re-run
NuGet restore."