[Core] Fix project in solution folder not reloaded
When a project inside a solution folder was modified outside the IDE
it was not reloaded. The Solution's ItemReloadRequired event was not
fired for projects inside a solution folder so the workspace was not
reloading them.

Fixes VSTS #669787 - Switching branch outside of IDE does not trigger
project reload
Fixes #6013 - Modifying .csproj file that is inside SolutionFolder
doesn't reload project