[NuGet] Fix PackageReference with no Version not shown in Packages folder
If the PackageReference, in a non-SDK project, had no Version
then it was not displayed in the Packages folder and a null reference
exception was logged. The PackagesFolderNode would try to find
the package to check if it was installed which is not possible
with a missing version and NuGet's VersionFolderPathResolver
would throw a null reference exception.