[Core] Reduce number of file watchers used with .NET Core projects
A file watcher was being created for each SDK style project. When a
solution contains a lot of SDK style projects (e.g. OrchardCore
contains 126 projects) this can then cause Mono's networking stack
to stop working. NuGet restore, check for updates would fail with
errors similar to:

SecureChannelFailure (Unable to write data to the transport connection:
The socket is not connected.)

Since the FileWatcherService has a maximum number file watchers the
project can use the common file watchers instead of creating its own.
This reduces the number of file watchers. For the OrchardCore solution
this reduces the number of file watchers down from 127 to 1.

There are still other NuGet restore problems with the OrchardCore
solution, but this solves one part.

Fixes VSTS #678422 - Unable to restore packages in OrchardCore app