[Ide] Improve Pasting into Solution pad
When pasting data into the solution pad, it might happen
that the currently selected node can not handle the data.
This is for example the case if a file is selected, the user
would expect the data to be pasted into the same folder.

This patch adds some logic to try pasing into the parent
node, if the currently selected node can not handle it.

Fixes VSTS #666943