[NUnit] Add configuration option to disable support assembly loading
The current logic for `IsSupportAssembly` doesn't take into account PackageReference binaries quite well since it cannot accurately determine with the availble information if such a reference will end up being copied to the output directory or not. This causes a lot of assemblies to be loaded in the running process mistakenly which in our case cause AssemblyLoad errors to happen.

This commit adds a workaround for now in the form of a project configuration option (set as a .csproj MSBuild property) to disable the current logic and essentially count on the fact that the necessary dependency closure for the tests are already being properly copied out by the build logic.