[NuGet] Move NuGet sdk resolver tests to NuGet addin
The NuGet sdk resolver tests rely on the MSBUILD_NUGET_PATH set
to point to the NuGet assemblies provided by the NuGet addin. The
NuGet sdk resolver test was failing in MonoDevelop.Core.Tests because
this was not being configured. The test has been moved to the NuGet
addin so the NuGet sdk resolver test will work from the command
line when running mdtool run-md-tests. The test works when run in the
IDE since the NuGet addin in the running IDE sets this environment
variable.