[NuGet] Show summary of NuGet restore errors
Creating an xUnit .NET Core test projet named 'xunit' would fail to
restore since there is a package reference cycle between the project
and the xunit NuGet package. Whilst the error message has been
improved in NuGet 4.7 to show there is a cycle this information is
buried in the Package Console output and all you would see was just
Restore failed. Now the error information is shown at the end of the
Package Console as a summary of the failures to make it easier to
see what failed. Now for the xunit project you will see the following
at the end of the Package Console output:

    Cycle detected.
      xunit -> xunit (>= 2.3.1).
    Restore failed.

Fixes VSTS #650078 - Restore failed for xunit.