[Ide] Fix solution pad not refreshing after pad prefs applied
On restoring the pad state on opening a solution sometimes the
tree view was not refreshed. A symptom of this was the Show All Files
preference being enabled or disabled in the display options context
menu but the tree view was displaying all files or not hiding the
files based on the setting. The problem was that only the selected
nodes were refreshed instead of all tree nodes. Switching to use
RefreshRoots, which is also called when the display options are
changed using the context menu, ensures the entire tree is refreshed.