[Core] Add workaround for Mono using incorrect uri for proxy
Using the HttpClient to make a non-secure http request through a proxy
that requires basic auth would result in the request failing to
authenticate. The problem is that Mono uses the request uri not the
proxy uri when asking the IDE for credentials. This results in the
ProxyCache not finding the cached credentials. As a workaround the
uri is checked to see if there is an associated proxy uri and if so
another attempt is made to get the credentials.

One example of this happening is in the Add Packages dialog where
the EntityFramework image was not being loaded since its image uri
is http://go.microsoft.com/fwlink/?LinkID=288859 and Mono uses this
uri when requesting credentials. https requests as expected, passing
the proxy uri when getting credentials from the ProxyCache.

https://github.com/mono/mono/issues/10622