[WebReferences] Support connecting through a proxy
Use the WebRequestHelper to provide proxy authentication support
when adding web references. Previously viewing a web reference
through a proxy would fail. The DiscoveryClientProtocol uses
HttpWebRequest and HttpWebResponse so the HttpClient cannot be used.
Instead a MonoDevelopDiscoveryClientProtocol class which derives
from DiscoveryClientProtocol is used. This new class uses the
WebRequestHelper when getting the web response so proxy auth is
now supported.