[WebReferences] Fix failure to view services requiring auth to view
With a web service that was protected behind basic auth it would
not be possible to view the service in the Web References dialog.
The problem was that the request was not being recreated when the
WebRequestHelper was re-trying the request with the user supplied
credentials. The MonoDevelopDiscoveryClientProtocol has been changed
to re-create the WebRequest instead of re-using the original request.
Also removed the AskCredentials dialog. The WebRequestHelper provides
its own credentials dialog so the AskCredentials dialpog is not
required.