[Mac] Handle basic auth protected VSTS NuGet package sources
Attempting to use a VSTS NuGet package source would prompt for
credentials but never be able to connect since the credentials were
not used in the request.

The NSUrlSessionHandler currently only supports NTLM responses.
However some WWW-Authenticate basic auth responses are not passed
to the NSUrlSessionHandlerDelegate's DidReceiveChallenge. Whilst
those that start with 'Basic' in the WWW-Authenticate header value are
passed to DidReceiveChallenge, those that start with something else,
such as Bearer, which VSTS NuGet package sources return, do not get
passed to DidReceiveChallenge. To handle this the response is checked
in the IDE and the Authorization http header is added if a credential
can be found and the request is retried again a single time. This
allows VSTS NuGet package sources