Port over MonoTODO analyzer to IOperation
Instead of binding a whole document, we wait for roslyn to do the
binding and only work with the given bound analysis.

This improves edit-time performance of analyzers, as they do not
try to bind symbols on every other keypress, but actually when
the code is compilable and method overload resolution is done, etc.