[Core] Try using a hashset to add item results
10 seconds of CPU time is spent in the All LINQ call which does a lot
of work to figure out whether an item is in the list (with the result
from CoreCompileDependsOn running containing a lot of items).

Try using a HashSet which would make this an O(1), not O(n). This
should take down OnGetSourceFiles time on Main.sln by 14s of CPU time.

Fixes VSTS #719932 - OnGetSourceFilesAsync can be optimized