[Debugger] Add Task api to DebugValueConverter
Added a GetValueAsync method which takes a CancellationToken and
returns a Task<T>. The default implementation calls the synchronous
GetValue method. Currently the async method is used when loading
images with the HttpClient. At some point the synchronous methods
should be deprecated and the debugger should use the async task
methods however some parts of the debugger api do not support tasks.