[Core] Fix incorrect DependentUpon metadata for Xamarin.Forms
The DependentUpon metadata for a .xaml.cs file in a subdirectory would
include the directory path instead of being just the .xaml file name.
This was become the include for the item included a backslash for the
directory separator. This backslash then caused the %(Filename)
MSBuild property to be incorrectly evaluated as the relative path
including the subdirectory, with the backslash character.