[NuGet] Allow verbose output when getting restore info from msbuild
Added an environment variable which will show the msbuild output
when the GenerateRestoreGraphFile target is run in the Package Console
window.

Also log a message when the restore info is being obtained from the
GenerateRestoreGraphFile target to the Package Console. If this
takes some time nothing was being displayed in the Package Console
window.