[NuGet] Improve solution restore performance
For solutions that contain many projects generating the restore
information for each project can be slow compared with getting
it in one step. The solution restore now uses similar approach to
what nuget.exe uses. MSBuild is run with a generated project file
that includes the PackageReference projects that are to be restored.
This allows the package spec information to be generated in one step
which is faster than using the msbuild host and getting the package
information for each project, when each project depends on several
other projects, such as with OrchardCore.

Timings: OrchardCore - git clean before each test

Package restore info - one project: 2 mins.
Package restore info - solution: 30 seconds