[Ide] Fix new folder not being selected in Solution pad
Sometimes when right clicking the project and selecting Add - New
Folder the project tree node is selected and made editable instead
of the new folder. The ContextMenuTreeView registers a tree view
SelectFunction in its OnButtonPressEvent which disables selection.
Sometimes there is no corresponding OnButtonReleaseEvent to undo
this SelectFunction so the TreeNodeNavigator cannot select the
new folder node that has been added. Now the TreeNodeNavigator
forces the SelectFunction to be reset before selecting a tree node.

Fixes VSTS #721437 - Can't focus on a folder and name it during
creation.