[NuGet] Modify reference in place on updating a NuGet package
On updating a NuGet package in a project that uses a packages.config
the old NuGet package is uninstalled and the new one is installed.
This removes the old references and adds new references. If the
references are in an ItemGroup with a condition then the new
reference can be added into a different ItemGroup if there are other
ItemGroups with references. To prevent this from happening the
changes to made to references are cached and not applied to the
project until the NuGet actions have all been run. This allows
a reference remove followed by an reference being added to be
converted into an update of the original reference's HintPath so its
location in the project file is not changed.

Fixes VSTS #697600 - Sometimes updating package changes its Reference
ItemGroup