[Core] Keep fully qualifed hint path on updating reference
Updating a NuGet package, where the project had been modified so the
original reference had a fully qualified hint path, would result in
a relative path used for the hint path on updating the reference.
Now if the original hint path was fully qualified then if the hint
path for the reference is changed it is saved using a fully qualified
path.

Fixes VSTS #697601 - Updating package reference adds relative path to
HintPath