[Core] Fix out of date reference information used by type system
On updating a NuGet package where only a hint path is changed for
a reference the type system would use the old reference information.
The problem was a race condition where the project modified event
is fired, the type system then gets the reference information from
the project builder whilst the project is being saved, after the
modifiedInMemory event is reset but before the project builder has
been refreshed. This resulted in the old reference information
being returned to the type system.