[Core] Fix old HintPath being saved
Saving the project after updating the HintPath worked the first
time but saving the project again the old HintPath would be used.
This was because the hint path changed flag was being reset. Once
the HintPath is modified it needs to be written to the MSBuildItem
on saving every time.