[Ide] Improve error handling when loading project templates
Failing to access the ~/.templateengine folder due to permissions
or not owning the files would result in the New Project dialog not
opening when trying to create a new project. Now a message dialog
is displayed showing the error before the New Project dialog is
opened.

Also moved the initialization of the templating engine out of the
constructor, to prevent the underlying error being hidden by a
TypeInitializationException. Initialization is now attempted each
time the templates are accessed in case the underlying error has
been fixed whilst the IDE is running.

Fixes VSTS #725161 - Failed to open New Project Dialog. Access to the
path "/Users/user/.templateengine/Visual Studio/7.0/..." is denied.