[NuGet] Cache package spec information
If a project is restored using the MSBuild host and it has project
dependencies then the package spec information is created for
these referenced projects. This package spec information is added
to the DependencyGraphContext. Previously only the project being
restored was added which could result in the GenerateRestoreGraphFile
target being run again for information that had already been
obtained.

There is still some overlap here if two projects share dependencies
but do not reference each other. This may still result in a restore
running for the same project dependencies. For projects that do not
have a large amount of dependencies this is OK and is still faster
than running MSBuild externally.