[Ide] Prevent duplicate solution in Solution pad
If the SolutionPad is initialized when the RootWorkspace's
FirstWorkspaceItemOpened event is fired a solution would be opened
twice in the Solution pad. When the SolutionPad is initialized it
adds the RootWorkspace Items to the tree view. When the RootWorkspace's
WorkspaceItemOpened is fired the SolutionPad then adds that item to
the tree view. To handle this a check is now made to ensure the
item opened is not already in the tree.

Fixes VSTS #736893 - Solution/project duplicated in project tree