[Ide] Reduce file watchers used by EditorConfigService
The EditorConfigService was creating a file watcher for .editorconfig
files in every project directory and all directories back to the
root directory. Creating too many CoreFX FileSystemWatchers can
result in http requests failing with an error similar to:

Error: SecureChannelFailure (Unable to write data to the transport
connection: The socket is not connected.)

To avoid this problem the FileWatcherService is used and only if
a .editorconfig file is found is a file watcher created for that
directory. If the .editorconfig file is in the project or solution
directory then this will result in a single file watcher created
for the solution, instead of multiple file watchers.

Fixes VSTS #738596 - Http requests do not work after the IDE has
been opened a while