[Ide] Fix new folder not selected after refresh
Collapsing a project folder node in the Solution window, then
selecting Refresh from the context menu for that project, then
selecting Add - New Folder, would result in the project node not
expanding and the new folder node was not selected.

Expanding the project node before adding the file to the project
fixes this problem.

Fixes VSTS #721437 - Can't focus on a folder and name it during
creation.