[Core] Fix project builder null ref after project reload
If the project is reloaded or disposed whilst an MSBuild target
is being run by the project builder this can result in the next
MSBuild target using the same project builder, but failing to
increase its reference count since it will be shutdown. Then
a null reference exception can be thrown if the build engine
is used since it will be set to null on disposing the builder.

Fixes VSTS #740214 - NuGet packages are not getting restored for
ASP.NET Core projects with error "System.NullReferenceException"