[Core] Handle builder shutting down in engine's GetRemoteProjectBuilder
Reworked the logic to handle a project builder shutting down inside
the RemoteBuildEngine's GetRemoteProjectBuilder instead of having
the project builder remove itself from the list of builders when
ReleaseReference is called. The build engine checks the return value
of AddReference and if this false, which means the builder is shutting
down, it is removed from the list of builders and a new one is
created.