[Core] Fix new folder disappearring in sdk project
Adding a new folder to a SDK style project would sometimes result
in the folder briefly appearring and then being removed. This was
because the file watcher generates a created and then a deleted
event for the new folder. The folder would then be removed from the
project and removed from the Solution window.