[NuGet] Add offline .NET Core project restore test
Tests for the regression in NuGet 4.9 so it is found earlier than
during a manual test. NuGet restore should not go online to download
packages for a .NET Core 2.1 project when the 2.1 sdk is installed.
The NuGet fallback folder should be used. This test disables all
online nuget package sources, changes the global packages cache to
a temporary directory, and then tries to restore a .NET Core 2.1
console project. With NuGet 4.9 this new test fails to restore
with the error:

Unable to resolve 'Microsoft.NETCore.App (>= 2.1.0)' for
'.NETCoreApp,Version=v2.1'.