Bump editor and vs-threading to bring in completion UI
The bump of vs-threading is necessary as the VS-Editor projects are compiled with a newer version of the library whose awaiters (part of the `AwaitExtensions` class) implement `ICriticalNotifyCompletion` which the async state machine generated by csc make use of by casting to the interface.

However because it's the version exported by MonoDevelop that's loaded in-process and that version used to be older it would create runtime failures because in that version the awaiters do not implement those interfaces.