[Ide] Make InfoBar API return a Control
Right now there is no way to remove an InfoBar programmatically once it
has been added. This can be a problem for addins that show an InfoBar.
For example, iOS and Android addins each show an infobar in certain
situations. If Android addin shows the infobar and the developer closes
the solution without hiding the infobar, then loads a different solution
without an Android project the Android infobar will re-appear.

If we return a Control for the addin to reference then the addin can
dispose of it when it's no longer needed.