[DotNetCore] Fix type resolution errors for signed projects
With two sdk projects, Lib1 and Lib2, where Lib1 has its output
assembly signed, then in Lib2 any use of any types from Lib1 would
show errors in the text editor after the solution was built. Errors
similar to:

The type 'Class1' exists in both 'Lib1, Version=0.0.0.0, Culture=Neutral, PublicKeyToken='
and 'Lib2, Version=1.0.0.0, Culture=neutral, PublicKeyToken='

This was because the generated AssemblyInfo.cs file which exists in
the obj directory was available to the type system so there was a
mismatch in the assembly information from the assembly .dll on disk
and the project information which resulted in Roslyn creating two
references. The generated AssemblyInfo.cs file is now returned to
the type system when it calls Project.GetSourceFilesAsync.

Fixes VSTS #753821 - Inline type resolution errors when project
references are public signed