[Windows] Fix documents getting closed on save
File.Replace internally renames the destination file to a temporary file.

When that happens, the FileSystemWatcher then takes that event and makes
the document point to the temporary file.

When done, the File.Replace code will delete the temporary file, causing
the document to close.

Circumvent that by using different IO events, which are possibly slower
but maintain correctness

Fixes VSTS #718767 - File close when save