CommandManager: fix KeyDown dispatching to native first responder views
If we have not handled a GdkEventKey, and we have a first responder that
is not the GdkQuartzView NSView hosting the GTK shell, dispatch out
to the responder's KeyDown directly and tell GTK we've handled it.

This fixes input in the new editor, and avoids extra work that would
otherwise happen inside GTK (which is incidentally still broken).