[Ide] Remove button not keyboard accessible in Edit References dialog
The Remote button which is used to remove the selected references
could not get keyboard focus. This is now enabled. Note that the Remove
button still does not get focus if it is clicked only if you tab to
it.

Fixes VSTS #750234 - Accessibility: Remove ‘X’ button is not accessible
through keyboard