[Ide] Fix contract in edit references dialog
The secondary text displayed for the selected references on the
right hand side of the dialog was hard to read when a list item
was selected. Now the main IDE secondary selection text style is used
when the list item is selected which has better contrast.

Fixes VSTS #750333 - Accessibility: Content under the ‘System’ below
the selected references is not visible in normal mode and same issue
is observed when we enable invert colors and grey scale