[Packaging] Fix language metadata accessibility in project options
Tabbing to the language combo box in project options - NuGet Package -
Metadata would not announce the label associated with the combo box
with Voice Over. The problem is that the combo box with an enabled
text entry does not work well with the accessibility library. To
fix this the combo box no longer allows free text entry. If no
language is selected then 'None' is displayed.

Fixes VSTS #753486 - Accessibility: NuGet Package Metadata: Voice
Over is not reading the label of the "Language" combo box.

Fixes VSTS #753489 - Usability: NuGet Package Metadata: Text area
inside the "Language" combo box does not serves any purpose and seems
not to be useful in this scenario.