[Packaging] Show language display name in project options for metadata
In project options - NuGet Package - Metadata the language display
name is shown in the language combo box. Previously the language
code (e.g. 'en-GB') was displayed. The display name is more user
friendly and also matches the behaviour of Visual Studio on Windows
when it shows package metadata for .NET Standard projects.