[Packaging] Fix voice over not reading warning label in project options
In project options - NuGet Package - Build the check box to enable creation
of a NuGet package at build time will show a warning label if there is no
associated package metadata defined for the project. This warning label was
not being read by voice over when the check box is selected. Now the
accessibility text is updated and includes the warning text if it is applies
so this is now read by voice over when the check box is selected.

Fixes VSTS #753475 - Accessibility: Voice Over is not reading the alert text
revealed after the "Create a NuGet Package" checkbox is checked.