[CSharpBinding] Use MenuButton for code page selection
The ComboBox with a TextEntry does not work well with accessibility
on the Mac. Voice Over will read it as a TextEntry or a pop-up button
in different scenarios. To avoid this the ComboBox has been replaced
with a TextEntry and a separate MenuButton. The list of code pages
is about 45 items so the context menu for the MenuButton fits on the
screen. However if the screensize is small then the menu will scroll.