Copy the net45 version of MS.VS.Composition to output. (#96)
* Copy the net45 version of MS.VS.Composition to output.

See bug https://github.com/mono/mono/issues/12461 for details. The net45 version has a workaround for .NET Framework bug that is still present on Mono. Later versions of the binary do not include the workaround and we still need it.

We ExcludeAssets="runtime" to not copy the default version of VSMef to output (net46).
We also add a manual reference to where the .dll is on disk and make it Private true so it gets copied to output.

* Reference net45 version of MS.VS.Composition manually.