FileSystemWatcher: Ensure proper disposal
When `FSW.OSX.FileSystemWatcher.DisposeAll` is called by `Runtime`, it
calls `Component.Dispose` on the FSWs. This did nothing because the
Dispose pattern was broken in the platform-specific FSW classes.
`Dispose(bool)` was not overridden; instead it was hidden by a
`protected internal` method that was only called when a
`FSW.FileSystemWatcher` was directly disposed.

Fix the Dispose pattern, so that calling `Dispose` on the
platform-specific FSW classes does what it should. And now
`FSW.FileSystemWatcher` simply calls `Dispose` on them as well.

Also make a small tweak in `StartRaisingEvents` to ensure cancellation
in case of errors.