Fixes VSTS Bug 763702: [Feedback] Using NUnit, typing inside Assert.That(tes pops up a Create method dialog that gets in the way
https://devdiv.visualstudio.com/DevDiv/_workitems/edit/763702

The old CompletionWindowManager hack doesn't work in this case - so I
introduced a pseudo-parameter (can't change that API or use it in any
sane way). A roslyn refactoring is requesting now the changes and that
happens during completion as well - that's why the insert windows are
popping up. This prevents that from happening.