[Core] Compare sdk names case insensitively in the sdk resolver
In general the DotNetMSBuildSdkResolver is not used, however if the
Sdk name being resolved differed in case (e.g. Microsoft.Net.Sdk
versus Microsoft.NET.Sdk) then the MonoDevelop sdk resolver would
fail to find a match. The main MSBuild sdk resolvers seem to do
directory/file checks when resolving sdk resolvers which would be
case insensitive if the drive was not case sensitive. To avoid this
problem the sdk names are checked ignoring the case.