[Core.Tests] Avoid NuGet sdk resolver causing test to crash
Running the SdkResolution class test to verify that the MSBuild
version is passed to the Sdk resolver was crashing the test with
the master branch. On the release-8.0 branch this test does not
crash. To avoid the crash we pass an invalid sdk version so the
NuGet sdk resolver does not attempt to resolve the sdk, which was
causing the crash. Since we are not interested in the NuGet sdk
resolver here, the DotNetMSBuildSdkResolver is the one that needs
the MSBuild version and occurs after the NuGet sdk resolver,
bypassing the NuGet sdk resolver is OK. Also the NuGet sdk resolver
works OK when testing with the full IDE - there is also a separate
NuGet sdk resolver test that verifies that it works.

The crash was some problem with NuGet which does not happen during
normal IDE operation:

NuGet.Configuration.NuGetPathContext doesn't implement interface
NuGet.Common.INuGetPathContext

  at <unknown> <0xffffffff>
  at NuGet.Packaging.FallbackPackagePathResolver:.ctor <0x00039>
  at NuGetAbstraction:GetSdkResult <0x0027a>
  at Microsoft.Build.NuGetSdkResolver.NuGetSdkResolver:Resolve <0x000aa>
  at MonoDevelop.Projects.MSBuild.SdkResolution:GetSdkPath <0x00380>
  at MonoDevelop.Core.SdkResolverTests:UnknownSdk_DotNetMSBuildSdkResolverDoesNotFatalReportError <0x0011a>