[Ide] Show error/warning directly inside the New Folder dialog
No longer using a separate message alert dialog for errors, such
as a directory already existing. Now if there is a problem with the
folder name a warning icon is displayed in the dialog and you can
hover this to see the warning text. The Add button is disabled if
there is a warning icon displayed.