[Ide] Add New Folder dialog for adding new folders (#7003)
* [Ide] Add New Folder dialog for adding new folders

Due to various problems with the Solution window when adding a new
folder directly into the tree and then editing the label, instead
a New Folder dialog is used when creating a new folder for a project.

Fixes VSTS #721437 Can't focus on a folder and name it during creation.

* [Ide] Show error/warning directly inside the New Folder dialog

No longer using a separate message alert dialog for errors, such
as a directory already existing. Now if there is a problem with the
folder name a warning icon is displayed in the dialog and you can
hover this to see the warning text. The Add button is disabled if
there is a warning icon displayed.